requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
passlib
python-jose[cryptography]